Introduce Tiers
Tiers handle whether or not a user is upgrading, depending on their
current plan. A plan has no knowledge of a tier, but it does have knowledge
of all the features it has and accepts an instance of a user in an instance
of a plan so the stuff that's enabled, disabled, or exceeded can be known.